requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
black==26.3.1
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
dj-rest-auth==7.2.0
Django==6.0.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
freezegun==1.5.5
isort==8.0.1
kombu==5.6.2
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.9.4
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytokens==0.4.1
redis==7.3.0
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.6.0
